+++ core-ng/src/main/java/core/framework/internal/log/DefaultLoggerFactory.java
@@ -1,5 +1,7 @@
 package core.framework.internal.log;
 
+import core.framework.log.LogLevel;
+import core.framework.log.LogLevels;
 import core.framework.util.Maps;
 import org.slf4j.ILoggerFactory;
 import org.slf4j.Logger;
@@ -11,6 +13,13 @@ import java.util.Map;
  */
 public final class DefaultLoggerFactory implements ILoggerFactory {
     private final Map<String, Logger> loggers = Maps.newConcurrentHashMap();
+    private final LogLevels.Entry[] infoLevels;
+    private final LogLevels.Entry[] traceLevels;
+
+    public DefaultLoggerFactory(LogLevels.Entry[] infoLevels, LogLevels.Entry[] traceLevels) {
+        this.infoLevels = infoLevels;
+        this.traceLevels = traceLevels;
+    }
 
     @Override
     public Logger getLogger(String name) {
@@ -22,22 +31,19 @@ public final class DefaultLoggerFactory implements ILoggerFactory {
     }
 
     private LogLevel infoLevel(String name) {
-        // kafka log info for every producer/consumer, to reduce verbosity
-        if (name.startsWith("org.apache.kafka.")) {
-            return LogLevel.WARN;
-        }
-        // refer to org.elasticsearch.nativeaccess.NativeAccessHolder, to emmit warning under integration-test env
-        if (name.startsWith("org.elasticsearch.nativeaccess.")) {
-            return LogLevel.ERROR;
+        for (LogLevels.Entry entry : infoLevels) {
+            if (name.startsWith(entry.prefix())) {
+                return entry.level();
+            }
         }
         return LogLevel.INFO;
     }
 
     private LogLevel traceLevel(String name) {
-        if (name.startsWith("org.elasticsearch")
-            || name.startsWith("org.mongodb")
-            || name.startsWith("org.xnio")) {
-            return LogLevel.INFO;
+        for (LogLevels.Entry entry : traceLevels) {
+            if (name.startsWith(entry.prefix())) {
+                return entry.level();
+            }
         }
         return LogLevel.DEBUG;
     }