+++ core-ng/src/main/java/core/framework/internal/log/DefaultLoggerServiceProvider.java
@@ -1,5 +1,6 @@
 package core.framework.internal.log;
 
+import core.framework.log.LogLevels;
 import org.slf4j.ILoggerFactory;
 import org.slf4j.IMarkerFactory;
 import org.slf4j.helpers.BasicMarkerFactory;
@@ -37,7 +38,7 @@ public class DefaultLoggerServiceProvider implements SLF4JServiceProvider {
 
     @Override
     public void initialize() {
-        loggerFactory = new DefaultLoggerFactory();
+        loggerFactory = LogLevels.createLoggerFactory();
         markerFactory = new BasicMarkerFactory();
         mdcAdapter = new NOPMDCAdapter();
     }