+++ core-ng/src/main/java/core/framework/log/LogLevel.java
@@ -1,4 +1,4 @@
-package core.framework.internal.log;
+package core.framework.log;
 
 /**
  * @author neo
@@ -6,7 +6,7 @@ package core.framework.internal.log;
 public enum LogLevel {
     ERROR(4), WARN(3), INFO(2), DEBUG(1);
 
-    final int value;
+    public final int value;
 
     LogLevel(int value) {
         this.value = value;