+++ core-ng/src/test/java/core/framework/internal/log/LogEventTest.java
@@ -1,5 +1,6 @@
 package core.framework.internal.log;
 
+import core.framework.log.LogLevel;
 import core.framework.log.Markers;
 import org.junit.jupiter.api.Test;